requirements.txt
torchvision>=0.8.1
pytorch-lightning>=1.0.8
pytorch-lightning-bolts>=0.2.5
wandb>=0.10.11
PyYAML>=5.3.1
scikit-learn>=0.23.2
efficientnet-pytorch>=0.7.0
